+++ CHANGELOG.md
@@ -22,6 +22,7 @@
   * Added support for AzureAD Application Authentication;
 * Modules
   * Added M365DSCAgent;
+  * Added M365DSCReport;
 * Metadata
   * Updated Microsoft.PowerApps.Administration.PowerShell to 2.0.57;
   * Updated MicrosoftTeams to 1.0.6;
@@ -30,7 +31,7 @@
   * Introduced dependency on the Microsoft.Graph.Authentication module;
   * Introduced dependency on the Microsoft.Graph.Identity.ConditionalAccess
     module;
-  * Introduced depencendy on the DSCParser module;
+  * Introduced dependency on the DSCParser module;
 
 ## 1.0.4.39